introduced another default partial implementation to get rid of the redundant save/load overrides
